Remove old TimeAdjustmentsPage
