--- FunTrophy.Web/Pages/Editor/NewTimeAdjustmentsPage.razor
@@ -1,67 +1,0 @@
-﻿@attribute [AuthorizeRoles(UserRoles.Admin, UserRoles.User)]
-@layout EditorLayout
-@page "/editeur/bonus"
-
-<PageTitle>Bonus</PageTitle>
-<h1>Editer les bonus</h1>
-
-<div class="p-2 col-12 col-lg-8 col-xl-6">
-    <select id="team-type" class="form-select" @onchange="OnSelectedCategoryChanged">
-        @if (Categories is null)
-        {
-            <option>Chargement...</option>
-        }
-        else
-        {
-            @foreach (var category in Categories)
-            {
-                <option value="@category.Id">@category.Name</option>
-            }
-        }
-    </select>
-</div>
-
-<table class="table table-striped align-middle">
-    <thead>
-        <tr>
-            <th scope="col">N°</th>
-            <th scope="col">Couleur</th>
-            <th scope="col">Equipe</th>
-            <th scope="col">Secondes</th>
-        </tr>
-    </thead>
-    <tbody>
-        @if (TimeAdjustments is null)
-        {
-            <tr>
-                <td colspan=4 class="text-center">
-                    <div class="spinner-border" role="status" />
-                </td>
-            </tr>
-        }
-        else if (!TimeAdjustments.Any())
-        {
-            <tr class="p-2">
-                <td colspan=3 class="text-center">Aucun bonus</td>
-            </tr>
-        }
-        else
-        {
-            @foreach (var bonus in TimeAdjustments)
-            {
-                <tr @key="@bonus">
-                    <td class="fit">@bonus.Team.Number</td>
-                    <td class="fit">
-                        <div class="team-color" style="border-color:@bonus.Team.Color.Code">
-                            <span class="@(HtmlHelper.GetFaIcon(bonus.Team.Type))"></span>
-                        </div>
-                    </td>
-                    <td>@bonus.Team.Name</td>
-                    <td>
-                        <input type="number" class="form-control" @bind-value="@bonus.Seconds" @onblur="(_) => OnValidateTimeAdjustment(bonus)" onfocus="this.select();" />
-                    </td>
-                </tr>
-            }
-        }
-    </tbody>
-</table>
