+++ FunTrophy.Web/Pages/Editor/TimeAdjustmentCategoriesPage.razor
@@ -52,7 +52,7 @@
             <label for="category-name" class="form-label">Nom</label>
             <input id="category-name" class="form-control" type="text" @bind-value="@addCategory.Name" placeholder="Nom de la catégorie" />
         </div>
-        <div class="p-2">
+        <div class="p-2 col-4">
             <label for="category-description" class="form-label">Description</label>
             <input id="category-description" class="form-control" type="text" @bind-value="@addCategory.Description" placeholder="Description" />
         </div>
